requirement.txt
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.193
botocore==1.12.193
certifi==2019.6.16
chardet==3.0.4
cssselect==1.0.3
docutils==0.14
et-xmlfile==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
gensim==3.8.0
idna==2.8
inflect==2.1.0
jdcal==1.4.1
jieba3k==0.35.1
jmespath==0.9.4
lxml==4.3.4
newsapi-python==0.2.5
newspaper3k==0.2.8
nltk==3.4.4
numpy==1.16.4
openpyxl==2.6.2
pandas==0.24.2
Pillow==6.1.0
python-dateutil==2.8.0
pytrends==4.6.0
pytz==2019.1
PyYAML==5.1.1
requests==2.21.0
requests-file==1.4.3
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
soupsieve==1.9.2
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.24.3
xlrd==1.2.0
